countryapi/serializers.py
from dataclasses import fields
from pyexpat import model
from rest_framework import serializers
from .models import Country
 
class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'