requirements.txt
asgiref==3.5.2
colorama==0.4.5
distlib==0.3.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
